example/kling-api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
requests==2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.2
starlette>=0.37.2
typing-extensions==4.8.0
fastmcp==2.5.1
mcp==1.9.1
httpx>=0.25.1
aiohttp==3.9.1